refactor(gateway): move OpenAPI auth types to core
